Stub out transaction controller

+++ src/Domain/Account.cs
@@ -4,7 +4,7 @@ namespace bounty.Domain
 {
     public class Account
     {
-        public int Id { get; set; }
+        public ushort Id { get; set; }
         public string Name { get; set; }
         public string Description { get; set; }
         public decimal StartingBalance { get; set; }
